+++ odoo/local-src/specific_crm/models/__init__.py
@@ -8,3 +8,4 @@ from . import crm_lead
 from . import survey
 from . import crm_team
 from . import project_task
+from . import partner