+++ HISTORY.rst
@@ -23,6 +23,9 @@ Latest (unreleased)
 **Features and Improvements**
 
 * Install sale_contract
+* Remove signatures on leads
+* update transitions checks in crm
+* Add NDA Leads in customer's form
 
 **Bugfixes**